split Landing
